moving test lib folder
