+++ test-js-lib/basic-election.js
