+++ test-js-lib/tiered-election.js
